// src/enums/user.ts
/* eslint-disable no-unused-vars */

// Enum User Roles
export enum ENUM_USER_ROLES {
  SUPER_ADMIN = 'super_admin',
  ADMIN = 'admin',
  STUDENT = 'student',
  FACULTY = 'faculty',
}

// Enum User Roles Short Code
export enum ENUM_USER_ROLES_SHORT_CODE {
  SUPER_ADMIN = 'SA',
  ADMIN = 'A',
  STUDENT = 'S',
  FACULTY = 'F',
}